Moved useEffect to relaunch from AppContainer to FolderTrustDialog
